pyproject.toml
[project]
name = "btc-news-sentiment-analysis"
version = "0.1.0"
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
  "pandas>=2.2",
  "numpy>=1.26",
  "pyyaml>=6.0",
  "pytz>=2024.1",
  "requests>=2.31",
  "feedparser>=6.0",
  "vaderSentiment>=3.3",
  "textblob>=0.18",
  "transformers>=4.42",
  "torch>=2.2",
  "safetensors>=0.4",
  "plotly>=5.22",
  "streamlit>=1.36",
  "pytest>=8.2"
]

[tool.pytest.ini_options]
pythonpath = ["."]
